Move check_user_in_room, check_user_in_room_or_world_readable, and check_can_change_room_list to EventAuthHandler.
